requirements.txt
jinja2
Pillow
mega.py
gdown